+++ tests/array/utils/gtest4dTest.h
@@ -6,17 +6,17 @@ extern "C" {
 
 #define SCALAR double
 
-const uint ARRAY_SIZE_X = 14;
-const uint ARRAY_SIZE_Y = 9;
-const uint ARRAY_SIZE_Z = 7;
-const uint ARRAY_SIZE_W = 6;
+const size_t ARRAY_SIZE_X = 14;
+const size_t ARRAY_SIZE_Y = 9;
+const size_t ARRAY_SIZE_Z = 7;
+const size_t ARRAY_SIZE_W = 6;
 
 class Array4dTest : public ::testing::Test {
 public:
-  uint IterAbsOffset(array4d::iterator iter) {
+  size_t IterAbsOffset(array4d::iterator iter) {
     return iter.i() + ARRAY_SIZE_X * iter.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * iter.k() + ARRAY_SIZE_X * ARRAY_SIZE_Y * ARRAY_SIZE_Z * iter.l();
   }
-  uint IterAbsOffset(array4d::const_iterator citer) {
+  size_t IterAbsOffset(array4d::const_iterator citer) {
     return citer.i() + ARRAY_SIZE_X * citer.j() + ARRAY_SIZE_X * ARRAY_SIZE_Y * citer.k() + ARRAY_SIZE_X * ARRAY_SIZE_Y * ARRAY_SIZE_Z * citer.l();
   }
 
@@ -50,8 +50,8 @@ protected:
   static array4d::const_pointer cptr, cptr2;
   static array4d::iterator iter, iter2;
   static array4d::const_iterator citer, citer2;
-  static uint offsetX, offsetY, offsetZ, offsetW;
-  static uint viewLenX, viewLenY, viewLenZ, viewLenW;
+  static size_t offsetX, offsetY, offsetZ, offsetW;
+  static size_t viewLenX, viewLenY, viewLenZ, viewLenW;
 };
 
 array4d Array4dTest::arr(ARRAY_SIZE_X, ARRAY_SIZE_Y, ARRAY_SIZE_Z, ARRAY_SIZE_W, ZFP_RATE_PARAM_BITS);
@@ -60,5 +60,5 @@ array4d::pointer Array4dTest::ptr, Array4dTest::ptr2;
 array4d::const_pointer Array4dTest::cptr, Array4dTest::cptr2;
 array4d::iterator Array4dTest::iter, Array4dTest::iter2;
 array4d::const_iterator Array4dTest::citer, Array4dTest::citer2;
-uint Array4dTest::offsetX, Array4dTest::offsetY, Array4dTest::offsetZ, Array4dTest::offsetW;
-uint Array4dTest::viewLenX, Array4dTest::viewLenY, Array4dTest::viewLenZ, Array4dTest::viewLenW;
+size_t Array4dTest::offsetX, Array4dTest::offsetY, Array4dTest::offsetZ, Array4dTest::offsetW;
+size_t Array4dTest::viewLenX, Array4dTest::viewLenY, Array4dTest::viewLenZ, Array4dTest::viewLenW;
